patch: When searching for offset, consider offset of previous hunk.
If the hunk doesn't apply to its intended place and doesn't apply to its
place + offset of the previous hunk, we search for any other placement
where the context matches.

We must pick the one that is closest to the intended place + offset of
the previous hunk, NOT the one that is closest to just the intended place.