Apply all changes to file at once
Since we adopted the behavior of patch that multiple hunks modifying a file
must not overlap, we don't really need to modify the file after every hunk.
Instead we can decide every hunk on top of the orignal content and then do
all the changes at once.

At this moment it is just a refactoring that simplifies some parts and
complicates others, but in future it will allow us to do the changes faster
and do some analysis on top of the original file while having the information
where all the hunks will be applied.