Recognize misordered hunks when applying, not when parsing
This now matches the patch behavior (if not completely, then at least more than before).
Patch doesn't care if the target lines of hunks appear out of order. It only cares about
the order in which the modifications would happen. So it is possible to have patch
with seemingly misordered hunks that applies fine (thanks to offets) and also the exact
opposite.