Skip file creation/delete on rollback if it failed on apply
This fixes abort in case two patches failed to apply and the later one failed
to delete/create file. When the later one was rolling-back to reach the state
where the earlier failed, it would panic because it was not able to do the
rollback. The correct thing to do is to not attempt the rollback, because since
it failed to apply, there are no changes.