Fix race when deleting empty directories
We used to check for empty directories and delete them right after a file was
deleted. When this is done by multiple threads at the same time, they may race.
For example one may delete a directory in which the other was just trying to
create new file. This has manifested as rare IO errors when saving modified
files.

Newly we collect the list of all directories that may be empty and check them
at the very end when all threads have written all their files.