Use jemalloc as global allocator
Jemalloc used to be the default allocator for Rust applications, but it was
changed to the system allocator in Rust 1.32.0.

While it is usually better to use the system-provided libraries, measuring
shows that jemalloc behaves much better in multi-threaded use.

Example benchmark on openSUSE Leap 15.0, 8 core Intel Xeon E5-1620 applying
44247 patches of SUSE's kernel-source branch SLE15-SP1:

  threads    system allocator    jemalloc
        1               5.2 s       5.1 s
        2               3.6 s       2.7 s
        4               3.0 s       1.6 s
        8               3.0 s       1.2 s

The amount of futex syscalls is also vastly different:
  system allocator: ~260000
  jemalloc:            ~400