Reduce the amount of stat calls
The `choose_filename_to_patch` function can return early if both old_filename
and new_filename are the same. No need to check if they exist. This prevents
lot of unnecessary stat calls. If the file does not exist, we will find out
when FileArena/MmapArena tries to open it.