Remove all interning
It turns out that interning all lines of files and patches was taking more time
than what was saved by it. Lazy interning was faster, but not as fast as
removing it all.

Instead of LineId, a unique id of the line, there is now directly &[u8], slice
of bytes from the file or patch stored in Arena.

Speed ups:
 * No interning at all. Patches are kept in their text form as they came from
   parser, files are only split by new lines.
 * Avoids allocation that was done when converting text patch to interned
   patch.
 * No need to lookup lines from interner when saving modified files.

Slow downs:
 * Comparing two lines means comparing their byte slices every time. Luckily
   the slices contain their length, which will often differ, so the comparison
   can bail early.
 * Files are represented as vectors of slices, so (on x86_64) 16 bytes per line
   instead of original 8 byte numbers (earlier even 4 byte numbers). So it
   consumes more memory and takes longer to move around when inserting/deleting
   from the middle of a file.