Get rid of util::search::Searcher
It was good when we needed to search array of u32, but it does not work
that well for searching array of [u8] slices. It relies on HashSet to
decide if given line is or isn't part of the needle. It was fast for u32,
but hashing the line is slower. It is better to just compare the lines
directly.