Handle failing to create a reject file in an unexisting directory
This was triggering a hang as the failure to create the reject file was
chained up to the main thread function which triggered an early exit
before calling wait() on the program's barrier. Blocking the rest of
threads forever.

We bypass creating the reject file altogether to mimic quilt's
behaviour: it doesn't create that .rej file yet correctly handles the
rest.

Signed-off-by: Nicolas Saenz Julienne <nsaenzjulienne@suse.de>