Fix applying patch when it's first in line
When pushing a specific patch rapidquilt validates it's not yet applied
by comparing its position in the series against the next patch to apply.
It fails to do so if the provided patch is the next in line in the
series, as it mistakenly includes that index as one of the ones to
avoid.

Signed-off-by: Nicolas Saenz Julienne <nsaenzjulienne@suse.de>