Never change current working directory
The current working directory is global state. Although it is in
fact OK to modify global state for a single run, it is known to
break parallel unit testing, and it may limit reusability of the
code in other projects.

Instead of changing the directory and using relative paths, make
sure to adjoin all relative paths to a base directory. AFAICS this
is an exhaustive list of all affected files:

- series file
- directory with patches (--patch-directory)
- files to be patched (input for patching)
- modified files (result of patching)
- quilt metadata (files under .pc)
  - applied-patches
  - per-patch backup files
- reject files

My goal was to keep user-visible relative (without the base
directory prefix).

Signed-off-by: Petr Tesarik <ptesarik@suse.com>