Pass a --directory argument to cmd::run() for testing
Do not change the current directory, because that breaks parallel
tests.

I am using OsStr to avoid dealing with non-UTF8 characters in
work_path (this will happen if they are contained in the TMPDIR
environment variable). However, it does not work just yet, because
getopts seem to be broken in that case...

Signed-off-by: Petr Tesarik <ptesarik@suse.com>