Do not panic if a file is truncated and the new name is bogus
When a patch fails, rapidquilt now always saves a backup file both
for the old and the new filename. Indeed, if they are identical,
the file gets written twice. However, the new name may be bogus if
a file is truncated to zero length, and rapidquilt then fails,
because the new file name cannot be found in modified_files.

Change this behaviour:

- Always save a backup under the old filename.
- Save a backup of the new filename if and only if
  applied_patch.file_patch.is_rename() is true. It is false for
  deleted and newly created files.

Also add a test case.

References: #20
Signed-off-by: Petr Tesarik <ptesarik@suse.com>