Simplify synchronization among parallel apply threads
The synchronization logic is quite complicated with all the
atomics, barriers and messages... IIUC this is the price paid for
maximizing thread utilization.

If we sacrifice a bit of efficiency, the logic can be greatly
simplified by splitting the job into two:

  1. Apply patches, finding the earliest one that fails.
  2. Save resulting files.

The earliest failed patch is not known until all threads are done
applying. AFAICS this synchronization point cannot be avoided, so
we may just as well terminate the first set of threads here,
provided that thread termination/creation overhead is negligible.

The only drawback is that no thread now starts rolling back until
all other threads are also done applying. However, it is not clear
what is better use of time, because there is no guarantee that
each thread gets its own dedicated CPU. If they happen to compete
over the same CPU, it might be better to terminate the thread,
letting another thread run towards the first goal (finding
rollback point).

Last but not least, there is no penalty if all patches apply
successfully.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>