Fix stderr capture while running the test suite
The capturing of stdout and stderr is implemented in
std::io::print_to(), which works well for the println! family of
macros, but not when you get the stderr handle explicitly.

The downside is that eprint!() requires a conversion to String,
which incurs some overhead. I think this is acceptable, because it
is done only once at the end of operation.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>