Correctly handle removal of empty files
Empty files should be removed if and only if the target file name
is /dev/null. GNU patch apparantly always takes the new name as
the target file name in this case.

Note that `quilt pop` removes all resulting empty files. I
consider it a quilt bug, but it might also be some sort of
compatibility.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>