Improve quilt_metadata test output
It's difficult to find the error messages in the output. Some
interesting data is in the Result error message, some is in the
captured output and some was sent directly to stderr, so it's
sprinkled somewhere amidst the progress messages...

Let's make sure everything is in the captured output, even if it
involves a lossy conversion from OsStr to String.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>