parser: Improve parsing of Git index lines
GNU patch actually parses the line and uses the data later. The
additional logic would have to be implemented yet, but as a first
step, parse the index lines.

Implementation note: An "SHA1" is any sequence of hex digit. In
particular, GNU patch does not check against the maximum length,
and neither does rapidquilt.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>