parser: Store SHA1 hashes in FilePatch
Since the hashes are now parsed, they should also be stored in a
FilePatch and output by the writer.

Note that git adds the file mode to the index line if and only if
the new mode is equal to the old mode. The "old mode" and "new
mode" lines are omitted in that case. Although the parsed mode
could be stored as old_permissions and new_permissions, this would
effectively cause that file permissions are forced to the new
value. But the presence of the mode means that the permissions
were not changed, so we should not touch them either. This is why
the optional mode in the index line is disregarded.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>