parser: Improve performance of discarded errors
The error messages in ParseError are all strings, because the
abstract Fail trait object is bound by 'static.

This was no big issue with nom-4, because the instance was created
only when nom::Err was converted to a ParseError using the From
trait. After migrating to nom-7, ParseError is now instantiated
when the error is encountered, requiring an immediate conversion
from &[u8] to String, even if the error is later discarded
(e.g. because a different parser alternative was taken).

The overhead can be avoided if ParseError contains a reference to
the input error location. There is a catch.  If only a reference
is stored, it must not outlive the referenced input, but the error
may have to outlive the parser input. To solve this, let's add a
StaticParseError type which is bound by 'static, but is
instantiated only when an actual error is returned by the parser.

References: #23
Signed-off-by: Petr Tesarik <ptesarik@suse.com>