parser: Rewrite parse_hunk_line using primitives
This allows to follow the error paths in more detail, making sure
that the only errors are BadLineInHunk and UnexpectedEndOfFile.
The latter is now returned when no (complete) hunk line is
available.

As a consequence, the error handling in parse_hunk() can be
greatly simplified.

The change also seems to reduce some overhead (measured by
valgrind's Ir units).

Signed-off-by: Petr Tesarik <ptesarik@suse.com>