parser: Rename parse_sha1 to parse_git_hash
A git OID may not always remain a SHA1. This field could also be
used to support other types of patches that store a file hash.

Just as important, I have added a comment about the differences
between git and GNU patch and the reason for writing a parser that
behaves slightly differently from both.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>