parser: Differentiate between missing hunk header and other errors
Both GNU patch and git apply recognize the beginning of a patch
hunk by checking whether a line starts with the marker ("@@ -").
If the rest of the line cannot be parsed, this is regarded as an
error ("patch: **** malformed patch at ").

OTOH rapidquilt does not detect such errors. If a line starts with
the marker but is not a valid hunk header, it gets ignored as
garbage. I believe that rapidquilt behaviour should get a bit
closer to that of GNU patch in this regard.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>