parser: Test invalid quoted filenames
It's not well defined what should happen if a filename starts with
a quote character (`"`) but is not a valid GNU-style filename:

- GNU patch ignores the filename
- git apply falls back to parsing it as unquoted

The git behaviour is more useful: Even if it is in fact a
malformed GNU-style filename, the path parsed as legacy will most
likely not exist, and users get a hint that there is something
wrong with the filename. Such issues are much harder to find if
the filename is silently ignored.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>