parser: Improve handling of unterminated quoted strings
The interface to parse_c_string() is currently somewhat weird. It
takes a pointer to the inner part of the string (i.e. after the
opening quote), expects a terminating quote, but will also parse
the string if it ends with a newline or EOF.

Sure, the closing quote is later checked by parse_filename_quoted,
turning a premature end of line or EOF into a Tag parse error, but
that is the consufing part.

Instead, parse the complete string in parse_c_string(), consuming
the opening and closing quotes. This allows to implement error
checking if the string is not properly terminated.

As a side effect, parse_filename_quoted() is no longer needed.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>