Rewrite patch parser without nom
While the nom parser is good and fast, it makes it sometimes
difficult to mimic the exact quirks that are present in existing
parsers.

The code still uses nom::ErrorKind nom::error::ParseError to make
this commit slightly smaller.

Oh, and as a matter of fact, I'm still fed up with the migration
efforts from nom-4 and want to dependencies.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>