parser: Introduce a NoMatch error
This error is meant to provide a low-overhead error code for
marking an alternative that does not match and similar cases. It
must be handled internally by the parser itself. It does not have
a corresponding StaticParseError; an attempt to create a
StaticParseError from ParseError::NoMatch would be in fact a logic
error, and the parser panicks in that case.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>