parser: Match first byte of metadata first
In an if-else block, Rust evaluates all conditions in the given
order, always trying to match the complete prefix. To recognize a
garbage line, all supported metadata prefixes must be tried...

By checking just the first byte, we can reduce this overhead.

Note that the effect was not measurable in wall time, but it
callgrind shows that the total Ir count has gone down by approx.
0.1 % for a 55000 kernel patch series on a dry run.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>