Do not use thread number to synchronize verbose messages
All verbose messages are currently logged by thread with ID 0.
This ensures that the message appears only once. However, there is
no guarantee that thread with ID 0 is the first one to reach the
corresponding stage.

Instead, let any thread print the message, but use an atomic bool
to ensure that the message is printed only once.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>