diagnostics: Use a simpler algorithm to find matching lines
While the current algorithm may find the best place for a failed
hunk, merely calculating the Levenshtein distance for all pairs of
lines from the hunk and from the file gets prohibitively expensive
for a hunk with a few hundred lines patching a file with a few
dozen thousand lines.

I don't think it is necessary to work on the character level. A
patch usually contains at least some lines that match exactly;
they can be used to find the best place.

We still compare each line of the hunk with each line of the file,
but we can at least omit the conversion to UTF-8.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>