diagnostics: Do not underflow starting file line number
If the best match for a hunk aligns with the beginning of a file
so that the first hunk line corresponds to a (non-existent) line
before the file beginning, then file_line overflows.

An illustration of the issue follows.

      FILE                   HUNK
                         [0] no-match line #1
                         [1] no-match line #2
  [0] something else     [2] no-match line #3
  [1] matching line #1   [3] matching line #1  <- start_line
  ...

In this case start_line corrsponds to hunk line at [3]. The
beginning of the hunk would align with file line [-2] (which does
not exist), but since file_line is an unsigned type, it overflows,
file line at [0] is skipped and the matching line is printed with
a bogus line number (e.g. 18446744073709551614).

Signed-off-by: Petr Tesarik <ptesarik@suse.com>