diagnostics: Do not attempt to access file lines after EOF
If the best match for a hunk aligns with the end of a file so that
the last hunk line corresponds to a (non-existent) line beyond
EOF, then rapidquilt panicks as it tries to use an invalid index
into modified_file.content.

An illustration of the issue follows.

        FILE                   HUNK
  ...
  [n-1] matching line    [n-3] matching line
                         [n-2] no-match line #1
                         [n-1] no-match line #2

In this case, next_file_line is advanced two lines beyond EOF.

While making this change, add some convenience variables (file_len
and hunk_len).

Signed-off-by: Petr Tesarik <ptesarik@suse.com>