diagnostics: Prefer matches that are close to the target line
If an equally good match can be found in multiple different places
of the file, the Dijkstra algorithm returns the first one. Instead,
pick the one that is closest to the target line number. This works
by increasing the cost of the initial step by the difference from
the target line as a fraction of total file length.

The maximum such penalty should be less than the cost of a differing
line, i.e. the "position cost" should be in the range [0; 1). To
allow keeping integer arithmetics, I'm multiplying these values by a
constant, so the cost of a line of difference becomes SCORE_SCALE,
and the position cost is then between 0 and SCORE_SCALE.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>