Preserve file permission for rollback and backup files
The permissions of a ModifiedFile are manipulated by
TextFilePatch::apply_internal(), either taking the permissions
specified by the patch, or inheriting the current permissions.

For newly created files, file permissions are initialized to
None. As long as no patch specifies permissions, file permissions
are not touched, and there is no issue. However, if a patch does
specify file permissions, the new value is stored in ModifiedFile,
and when that patch is rolled back, that value is also propagated
back. This is wrong. If a file had no explicit permissions, it
should be restored with None permissions during rollback.

Fixes: #26
Signed-off-by: Petr Tesarik <ptesarik@suse.com>