Remove backup files if create() fails with PermissionDenied
If the same target file is patched more than once by the same patch file,
then each patch is treated more or less the same as if it were contained in
a separate patch file. However, the backup file path is the same for all
such patches, so the backup file may be written more than once. Since the
backup file preserves the original file's permissions, such repeated write
may try to open a read-only file for writing. This fails with -EACCES,
which translates to a PermissionDenied error kind in Rust.

Remove the target file and retry the create() if the first attempt lacked
the necessary privileges to complete.

Fixes: #32
Signed-off-by: Petr Tesarik <ptesarik@suse.com>