Do not track fuzz/Cargo.lock
Similar to the Cargo.lock in the root directory, this file can be
re-generated as necessary, and we don't want to lock into a specific
version of a crate, increasing churn whenever it needs to be updated to fix
an upstream bug.

See also commit 1e3e7c8dbb79 ("Remove Cargo.lock").

Signed-off-by: Petr Tesarik <ptesarik@suse.com>