libpatch: Fix off-by-one frozen limit
After applying a hunk, the last frozen line is updated to the index of the
first unmodified line rather than the last modified one.

Rather than subtracting one in the calculation, change the semantic and
track the index of the first modifiable line. The added benefit is that
this index is never negative (which be useful later).

Signed-off-by: Petr Tesarik <ptesarik@suse.com>