libpatch: Add a `movable` parameter to try_apply_hunk()
Let the caller determine whether a hunk can be applied with an offset, or
not. This is needed in try_apply_hunk() to find all possible target lines.

Hunk movability depends on the apply mode and hunk position. These
conditions either depend on the caller or are already checked in the
caller.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>