libpatch: De-duplicate applying a hunk to the modified file
Move the common code to HunkApplyReport and call the new function commit().

The idea behind HunkApplyReport is that it provides a detailed report of
what exactly was patched where. However, this is a two-stage process:

  1. Determine the location of each hunk.
  2. Apply the changes to the modified file.

The first stage is performed in try_apply_hunk(). This is used to create a
new report. The second stage was open-coded in apply_modify() and
rollback_modify(). It's the second stage that is de-duplicated by this
patch.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>