diagnostics: Treat newline as space
Do not consider newlines in the hunk and/or hunks. This change may improve
hunk matching where a file does not end with a newline.

Note that compare_ignore_space() may have iterated through all bytes in the
first slice, while there are still unprocessed spaces in the other slice.
Make sure that the function returns true in that case.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>