libpatch: Always store the file header in TextPatch
Always store the header slice in TextPatch. This has been controlled by a
parameter, but then the parser should not warn about possible hunk headers
in the leading garbage (the "file header") and the logic relies on that
parameter (by mistake).

Do not bother adding yet another parameter, because the cost of saving the
header slice is near to zero. It's merely a slice of the input, and its
lifetime is the same as that of the parsed hunks.

Fixes: #39
Signed-off-by: Petr Tesarik <ptesarik@suse.com>