diagnostics: Improve matching of hunks with modified lines
Currently, the cost of an arc in the match graph is the sum of lines
deleted and inserted. Since modified lines appear as both inserted and
deleted, their cost is effectively double the cost of an inserted or
deleted line.

Reduce the cost of modified lines by taking the maximum of inserted and
deleted lines instead.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>