diagnostics: Fix analysis output with empty lines
Rescan the best path for matching lines. Since (almost) empty lines were
treated as non-match, such matching pairs are missing from the best path.

Commit a94d53575ad7 ("diagnostics: Ignore empty lines") tried to address
the issue by adding a check at the beginning of the print loop, but that is
insufficient, because matching empty lines may appear anywhere in the hunk,
not just at the beginning. Now that we fix the issue properly, that old
attempt can be removed.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>